# Conflicts:
#	doc/pkg-src.stamp
#	doc/pkg/loft_bg.wasm
#	src/parser/definitions.rs
#	src/parser/mod.rs

+++ tests/testing.rs
@@ -570,7 +570,7 @@ impl Test {
             if let Value::Int(_) = self.result {
                 Type::Integer(IntegerSpec {
                     min: i32::MIN,
-                    max: i32::MAX as u32,
+                    max: i64::from(i32::MAX),
                     not_null: false,
                     forced_size: None,
                 })
